Improve error message locations for function calls and call arguments (#6009)
### What was changed?

Added a method for updating test expect files based on CI output. This
is contained in the C# project `Scripts`

#### Call arguments
Errors that relate to call arguments are reported on the argument,
instead of on the `(`.

Example:
```dafny
r := new C<char>.Orig('h'); // error incorrect argument type for constructor 
                            // in-parameter 'x' (expected X, found char)
                      ^ new location
                 ^ old location
                     
```

#### Function calls
Errors that relate to function or method calls are now consistently
reported on the `(`. This was already done for method calls, while for
function calls the error was reported on the center of the callee
expression. It's better to report on the `(`, to distinguish from errors
in computing the callee, which are reported on the center of the callee
expression.

Example:
```dafny
ghost const objs: set<object> := getObjs()
                                        ^ new location
                                 ^ old location
// Error: insufficient reads clause to invoke function
```

#### Assertions
Currently, errors about assertions (assert/ensure/invariant) are
reported on the center of the condition. Since this PR changes the
center of expressions that call functions, the locations for 'assertion
might not hold' can change as well. Example:
```dafny
assert !Even(N(17));
        ^^^^|^^^^^^ new origin
        |^^^ previous origin
// Error: assertion might not hold
```

Another example:
```dafny
ensures Pos(UpIList(n))
           ^ new location
        ^ old location
// Related location: this is the postcondition that could not be proved
```

### How has this been tested?
- Updated existing tests

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>