Fix: Explicit assertions no longer considered implicit for code actions (#6030)
Fixes #6025 
### What was changed?
Explicit assertions (which as of today are prefixed with ensures and
assert) are no longer suggested in code actions where the same
expression fails. This was redundant and useless.

### How has this been tested?
I added a test for the language server in the assertion case, verified
that this test did not pass before the change, and the test passes after
the change.
Since ensures are also explicit assertions, I added a test for them as
well. However, the test does not work if the ensures is not on the same
line as the opening curly brace. I don't know why, but for now I'm
writing the test in a way that was surely failing before and now passes.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>