Feat code actions forall calc (#6044)
### What was changed?
First, for implicit assertions, if they are surrounded by a declaration
that supports by clause (like AssertStmt or CallStmt), the by clause is
preferred to inserting the assertion before.

![InsertVarByStatement](https://github.com/user-attachments/assets/add8e4c4-f4e1-4818-9374-0c69dd10e387)
And that works also for well-formedness, since well-formedness is now
proved by by-blocks (thanks @keyboardDrummer for that, and @alex-chew
for creating Dafny expressions for well-formedness)

![InsertAssertByPreconditionStatement](https://github.com/user-attachments/assets/7b4fe592-abf6-4b31-9ee3-3cfc75749284)
Moreover, if a by clause already exist, the code action will insert into
it.

![InsertAssertExistingByPreconditionStatement](https://github.com/user-attachments/assets/bce3c787-1655-410d-9dc6-52fe924a976c)

Second, assertion of foralls now support a proof sketch by inserting
forall statement

![InsertForallStatement](https://github.com/user-attachments/assets/c411570d-8f33-463f-ab7a-af093a594a31)

Third, assertions of equalities or equivalence now support a proof
sketch by inserting a calc statement.

![InsertCalcStatement](https://github.com/user-attachments/assets/fb76312a-51b5-4325-a0cd-457ef5dc444e)

![InsertCalcEquivStatement](https://github.com/user-attachments/assets/68d12a15-684a-4f97-bd66-fa86ca8ad348)

The code for inserting such proof sketches has been refactored so that
it will be even more easy to extend in the future.

### How has this been tested?
- Tests were updated for the insertion point (var....by) when possible
and insert there if it already exist.
- Tests were added for foralls, equalities and equivalences, and also
new insertion points

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>