Fix: Generated Rust docstring no longer crashes cargo doc Fixes #6084
Previously, Dafny docstring was literally transformed to Rust docstring. However, Rust docstring follows a markdown syntax where non-labelled code blocks are interpreted and checked as Rust code when running `cargo doc`.
Rather than creating a flag to avoid the docstring, this PR does a lightweight parsing on the docstring to ensures that:
* In docstrings, Code blocks without label are written using the "dafny" label so that they are not interpreted as Rust
* In docstrings, text that has a certain level of indentation has its first space replaced by a vertical bar so that we make sure it's not interpreted as Rust code.

I added Dafny tests for the new function that converts docstrings, and also an integration test that verifies that 'cargo doc' works on the already existing but modified docstring.dfy test.