Fix: Generated Rust docstring no longer crashes cargo doc (#6085)
Fixes #6084
### What was changed?

Previously, Dafny docstring was literally transformed to Rust docstring.
However, Rust docstring follows a markdown syntax where non-labelled
code blocks are interpreted and checked as Rust code when running `cargo
doc`. Rather than creating a flag to avoid the docstring, this PR does a
lightweight parsing on the docstring to ensures that:
* In docstrings, code blocks without label are written using the "dafny"
label so that they are not interpreted as Rust
* In docstrings, code blocks with the `rs` are written without this
label to ensure it's possible to write tested Rust docstring that can be
executed with `cargo test --doc`
* In docstrings, text that has a certain level of indentation has its
first space replaced by a vertical bar so that we make sure it's not
interpreted as Rust code.

### How has this been tested?

I added Dafny tests for the new function that converts docstrings, and
also an integration test that verifies that 'cargo doc' works on the
already existing but modified docstring.dfy test.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>