Chore: Not generate boogie translation if not verifying (#6067)
### What was changed?
I ensure the `Options.DafnyVerify` is set to false if `--no-verify` is
provided, except if `--bprint` is set

### How has this been tested?
A new driver test verifies the new behavior. Without the test, I got it
wrong, so I'm confident that the test ensures the behavior is correct.
I don't think there is another way to test that the Boogie translation
does not happen when it won't actually verify;

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>