Opaque block fresh soundness (#6066)
Fixed #6060 

### What was changed?
Always introduce a new heap variable after an opaque block. I think if
no reference 'escapes' the opaque block, then the new heap variable is
not needed, and it would be better not to create it, but I couldn't find
the machinery to check for the escaping.

### How has this been tested?
Added a test for the soundness bug

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>