Add serialization framework (#6112)
Changes to enable reading serialized Dafny sources

### What was changed?
- Create a script command `generate-deserializer` which generates a
`Deserializer` class that takes a `IDecoder` and returns an instance of
the Dafny AST, which is in a post-parsing state.
- Enable the `dafny` CLI to consume `.dbin` files that contains
serialized Dafny code, and create an option `--input-format` that allows
the `--stdin` command to accept serialized input.
- Create a script command `generate-parsed-ast` which create a version
of the Dafny AST that only contains fields which filled in through
source code. This is useful for other languages to generate their
version of the Dafny post-parsing AST.
- Create a script command `source-to-binary` that takes a Dafny program
and outputs its serialized format, based on the C# types generated by
`generate-parsed-ast`. This command is used for testing both of the
above code generation commands.

To enable testing the above commands using a minimal "assert false"
Dafny program, the following changes were made as well:
- Add nullable annotations to the Dafny AST in some places, together
with `#nullable enable` annotations in this files, and fixes for the
nullability warnings in those files.
- Add `ParseConstructorAttribute` to some constructors in the Dafny AST
- Remove some fields from the Dafny AST that always held the same value.
- Rename certain fields/properties and constructor parameters, so the
names match

### Future work
- Currently the 'interface definition language' is C# types. We can
still switch to any of the existing IDL languages in a separate PR. I
think most of the code would remain the same. However, I'm not convinced
it makes sense to use another IDL so I won't further investigate that,.
- The serialization format is schemaless, and the code is setup so that
the format can be easily changed by providing a different implementation
of `IDecoder`. Currently there's only a text based format, but it would
be easy to add a byte based one. The text based format could be improved
with newlines and indentation so it's easier to read.

### Caveats
- Given an abstract type like `Expression`, the generator only visits
subtypes that have a constructor annotated with `[ParseConstructor]`,
which is intentionally missing for many types right now. Because it is
missing for those types, not the entire Dafny AST is generated, which
means we can work more incrementally.
- The generators require that each parameter of a constructor can be
mapped to a field or property, and fail otherwise. This PR does not
update CI so that the generators are run, so the AST could become out of
date after this PR is merged.

### How has this been tested?
- Added `inputFormat.dfy` CLI test, which takes a Dafny file, serializes
it to a file, deserializes it from that file, and then verifies it.
- TODO Add CI checks to ensure the generated code is up-to-date

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>