Fixes: Rust supports overriding traits (#5948)
### Description
The PR https://github.com/dafny-lang/dafny/pull/5940 defined a
preliminary support for value traits.
This PR fixes everything that encompasses value traits and newtypes.
Here is the list of fixes.
* Fix to support functions declared in a trait and implemented in
another trait
* Fix to support for equality in type parameters and conditional
equality support for datatypes.
* Fix to support constants in datatypes
* Fix to support constants in traits
* Fix to support back/forth conversions between &self and the underlying
type
* Fix to ensure we don't support coercions where it's not possible (that
caused compilation errors for Rust and it's behind the discovery of
soundness issue #5972)
* Fix support of equality for newtypes (that caused compilation errors
for Rust and it's behind the discovery of soundness issue #5978)
* Fix to support type bounds in type parameters
* Fix to support newtypes on booleans
* Fix to support correct borrowing of type arguments in methods when the
method instantiate a trait method which always borrows the arguments
with the type parameter
* Fix to support static constants in datatypes, including type
parameters
* Fix to support type parameters in static functions
* Fix to ensure type parameter equality is correctly generated
* Fix to support downcast from one value trait to another value trait of
the same hierarchy (needed to change the resolver to record all such
traits)
* Fix to support the enclosing module name in externs

Also some enhancements
* Unification of extern semantics: Remove the error message about
`{:extern}` on types requiring a hint when it can simply take the Dafny
name, like for modules and classes.
* `DafnyCodeGenerator.GenType` now caches its result to avoid
exponential computations


I also updated the documentation regarding the use of externs

### How has this been tested?
* All the tests of comp/rust/traits-datatypes.dfy and comp/rust/small/*
were not passing before this PR, and now they are.
* The way to downcast traits from one to the other is tested in the
test/mod.rs of the runtime.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>