chore(Python): Improve slicing performance (#6042)
Add a new class `Slice`. This is a lazy slice for `Seq`s. The elements
of a `Seq` are immutable, so slices of a `Seq` can refer to the original
`Seq`'s elements, and lazily calculate access based on the slice
indices. Before, a slice of a `Seq` would copy all elements by value
from the original to the new `Seq`.
  * Slicing a `Seq` now takes constant time.
  * Slicing a `Slice` takes constant time.
  * Getting a single element from a `Slice` takes constant time.
  * Getting the length of a `Slice` takes constant time.
  * Enumerating the elements in a `Slice` takes linear time.

---------

Co-authored-by: Fabio Madge <fmadge@amazon.com>