feat: Support using --standard-libraries with  --enforce-determinism (#6137)
### What was changed?
Build the standard libraries with `enforce-determinism = true` in
general, so consumers can use that option too. This is especially
important now since it turns out Rust code generation requires this
option.

This was blocked by the non-deterministic `var x :| x in s;` statement
in `Seq.SetToSeq`, so I've also removed that just as we did previously
from `dafny-lang/libraries`:
https://github.com/dafny-lang/libraries/pull/158#pullrequestreview-2249274198.
Unfortunately a breaking change, but done in the name of soundness, and
the implementation was O(n^2) and not suitable for most production use
anyway.


### How has this been tested?
Also updated standard libraries tests to specify `enforce-determinism =
true`.