Merge `RangeNode` and `NodeWithComputedRange` (#6148)
### What was changed?
- Previously, AST nodes had to choose whether to inherit from
`RangeNode` or `NodeWithComputedRange`, which in combination with type
hierarchy was rather inflexible. For example, all expressions inherited
from `NodeWithComputedRange`, and all statements from `RangeNode`. This
PR changes the mechanism so that when constructing an AST node, the
caller can decide whether it's providing the node's range, or whether it
should be computed from the node's children.
- Concretely, an `IOrigin` now no longer has a `Token StartToken` and
`Token EndToken` but a `TokenRange? EntireRange`. If `EntireRange` is
set, then this determines the range of the owning `INode`, and otherwise
the node's children do it. AST nodes now generally inherit from
`NodeWithOrigin`. There's also a new type `NodeWithoutOrigin` which is
an orthogonal change.
- This change allows for more often letting the `EntireRange` be
computed based on children, which allows writing less code. This is also
useful for users of Dafny's deserializer, since they have to provide
less token information.
- AST nodes that have no syntactic component but only wrap other nodes,
like `Specification`, can inherit from `NodeWithoutOrigin`. This way,
downstream tools can instantiate them without having to pass any source
location information.
- An IOrigin now has not a `Token Center` but a `TokenRange
ReportingRange`, so that the reporting range can be more than a single
token, which is useful for example for having `-1` as the reporting
range, which has two tokens.
- This change is particularly useful in combination with
https://github.com/dafny-lang/dafny/pull/6136, which is useful for
downstream tools consuming Dafny's output
- This change enables a follow-up PR that streamlines reporting across
the IDE and CLI, because with the `Center` now being a pair of tokens
instead of a single one, it is more suited for the IDE to be used for
reporting.

### How has this been tested?
- Some existing tests were updated. Please inspect them carefully.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>