Prepare for making new resolver the default (#5653)
This PR prepares for making the new resolver the default (i.e., making
`--type-system-refresh --general-traits=datatype --general-newtypes` the
default), but it does not actually change the default.

* It fixes many bugs in the new resolver and makes other improvements
(tighter inference, better error messages, etc.)
* It changes the test suite to run the new resolver, except in a few
dozen test cases that explicitly invoke the legacy resolver. (Among
those exceptions, many will be moved to use the new resolver in due
time, but for now it seems better to get these bug fixes in rather than
delaying this PR.) By default, tests that use
`%testDafnyForEachResolver` and `%testDafnyForEachCompiler` use both the
old and new resolver.
* It changes the building of the Dafny runtime, Dafny standard library,
`AutoExtern`, and documentation to use the new resolver.

This PR does not change the default resolver. So, users of the old
resolver are mostly unaffected, and users of the new resolver obtain the
PR's bug fixes.

## Commnents for reviewers

* The behavior of the resolver requires changes in some Dafny code,
which affects the test suite:
- In the new type system, a trait is not a reference type unless it's
declared with `extends object`.
- The new resolver infers all bound variables to be of base types (like
`int`), never of a subset type (like `nat`).
- In some places, specific type parameters have to be given explicitly
(usually because the type needs to be a subset type--maybe these can be
improved in the future).
- The old type system would automatically downcast (from a trait to a
class). The new resolver requires an explicit cast for this (as does C#
and Java, for example).

* Sometimes, the tests show a different number of things verified.
That's because the new or old type system takes care of all the proof
obligations.


<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>

---------

Co-authored-by: Remy Willems <rwillems@amazon.com>