Add compilation of multi-dimensional arrays to IR (#6140)
### What was changed?
Added support for the compilation of multidimensional arrays to the
Dafny IR (with help from @fabiomadge; thanks!); it was already part of
the AST for the IR, and the compilation to Rust seems to be implemented
already as well.

### How has this been tested?
Added a new test (`comp/rust/array2d.dfy`), which tests this feature by
compiling multidimensional arrays in the IR to Rust.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>

---------

Co-authored-by: Fabio Madge <fmadge@amazon.com>