fix: Generate alloc consequence axiom only for functions that read the heap (#6166)
Previously, Dafny generated axioms of the form

``` dafny
// alloc consequence axiom for _module.__default.On
axiom (forall $Heap: Heap, n#0: ref, f#0: int :: 
  { $IsAlloc(_module.__default.On(n#0, f#0), Tclass._module.Referrer(), $Heap) } 
  _module.__default.On#canCall(n#0, f#0) && $IsGoodHeap($Heap)
     ==> $IsAlloc(_module.__default.On(n#0, f#0), Tclass._module.Referrer(), $Heap));
```

for any function whose result type may involve references. This is not
sound, since it fails to connect the parameters with the `$Heap`.

The axiom would be sound if the `$Heap` parameter is part of the
function and/or its CanCall predicate. In the situation discovered in
#6164, the function (named `On`) does not have such a `$Heap` parameter.
This PR makes sure that, in those cases, the axiom's antecedent will
instead include explicit `isAlloc` predicates that connect the parameter
with the heap parameter.

Fixes #6164

### Note for reviewers

Previously, the local variables `ante` and `anteIsAlloc` in method
`AddFunctionConsequenceAxiom` were updated in various loops. But then
those built-up values were replaced with `useViaCanCall`. This PR cleans
that up. The PR also removed the unused local variable `pre` in that
method.


<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>