Chore: Fix compliancy of language server with LSP (#6174)
Fixes https://github.com/dafny-lang/dafny/issues/6172. I ensured that
two notification messages' parameters that are disregarded anyway in the
VSCode extension use the proper LSP semantics of being wrapped with an
array or being an object, so that the LSP is compatible with Helix for
example. I tested locally that this change does not break the VSCode
extension, not sure how to test it furthermore.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>