Optimize sequence equality comparisons in Go runtime (#6175)
### What was changed?
Sequence comparisons (`a == b`, `a < b`, etc) in Go are currently
implemented via code translated from the `dafnyRuntime.dfy` code. It's
quite slow because comparing each element uses `AreEqual` which operates
on `interface {}` values and has to use reflection. This PR replaces the
Dafny `EqualUpTo` method with an extern that defers to comparing slices
directly for a couple of types where this is known to be sound (bytes
and characters for now).

Also addressed a CI gap where the newer copy of the Go runtime was not being checked.

### How has this been tested?
Old Go runtime well-covered by existing tests. The new one is not (see
#6179), but was tested manually in consuming Dafny projects.

---------

Co-authored-by: Robin Salkeld <salkeldr@amazon.com>