Fix: Bitvector shift in `--type-system-refresh` (#6197)
Fixes #6196 

I updated the tests to ensure it was breaking and that this PR fixes the
test.
Basically, a test was testing if the type was a bitvector type, but this
type was a synonym type with the new resolver so that needed to be
updated.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>