Fix 6212 dafny to rust compilation issue with conversions (#6213)
Fixes #6212 
Previously, multiset indexing to retrieve the multiplicity wasn't
supported in the Dafny-to-Rust compiler, but it worked seamlessly for
multisets of integers because the compiler was treating it silently as a
`map<int, int>`.
Added proper support for multisets of any types.

### How has this been tested?
Test added

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>