Feat: `Std.FileIO.ReadFile` and `WriteFile` (#6198)
This PR adds an opinionated way to write Dafny strings to files and read
them back using UTF-8 encoding.
With this PR, it becomes finally possible to create command-line
utilities that read a file and process it in an easy way.

### What was changed?

- Added `FileIO.ReadUTF8FromFile` and `FileIO.WriteUTF8ToFile` so that
it's easier to read/write strings to text files within Dafny.
- `FromUTF8Checked`, `FromUTF16Checked` and
`DecodeCodeUnitSequenceChecked` all now return a `Result` instead of an
`Option` so that the error message is clearer.

### Example

For example, here is a tool that displays the content of a file
```
import opened Std.FileIO
method Main(args: seq<string>) {
  expect |args| >= 2;
  var result :- expect ReadUTF8FromFile(args[1]);
  print result.value;  
}
```

Usage:
```
dafny run --standard-libraries cat.dfy cat.dfy
```
It prints itself.


### How has this been tested?
I added tests to read and to write strings similar to the existing tests
to read and write UFT-8 characters.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>