Just run nightly build on 4.11 (#6245)
### What was changed?
Previous attempts to run the nightly build on both master and the next
release's branch have failed because they both attempt to upload the
same artifacts:
https://github.com/dafny-lang/dafny/actions/runs/15023604141/job/42218715250#step:29:26

For now it's adequate to only run the deep tests on the release branch.
Deep test regressions could happen on master in this state, but those
will be caught after 4.11 is released and we switch the nightly build
back to it.

Also deleted test-report.yml as it has not run in years, and it's
presence confused me as I tried to figure out what was broken. :)

### How has this been tested?
Manual run of workflow from source branch:
https://github.com/dafny-lang/dafny/actions/runs/15146440803

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>