Feat memory locations locals (#6242)
### What was changed?

This PR adds the ability to refer to local fields as memory locations
when the flag `--referrers` is set. For example, here is a summary of
what can be done
```dafny
method Test(i: int) returns (r: int)
 decreases *
{
  var j := 1;
  assert locals`i != locals`j; // OK
  var r := Test(i - 1) by {
    assert Test`i != locals`i; // OK, recursive call does not have the same memory location.
   }
}
```

There is no notion yet of tracking here, only the ability to refer to
the field of local variables. There is no dereferencing since these
local variables live not on the heap but the unique stack object,
represented as the `locals` variable.

For this PR, I had to solve two problems
* Input parameters can be referred both in the body of a method and in
the requires of a recursive call. Hence, to assert in Boogie that their
fields are different, we need a different encoding than object fields
(which are always unique).
* I first explored having two separate fields. However, these fields
should coincide at a call site when trying to satisfy the requires,
which would lead to contradictions.
* I then explored other solutions, such as checking the requires
manually so that I can do substitutions on field declarations, attaching
a depth or LayerType, make the stack a different object between a caller
and callee. I realized I wanted to keep the stack as a single "object"
(because later on, we will need uniqueness properties of that object),
and not change the encoding of satisfying requires.
* The solution I retained was to use a depth as an integer or layerType.
However, Boogie cannot prove `$LS($LZ) != $LZ)` so I went the route of
integers. Under the `--referrers` flag, every method and constructor's
first parameter is `depth: int` in Boogie, and that depth is increased
at every call. This depth is not exposed at the Dafny level (but could
one day?), but makes everything sound and practical.
* Referring to the input parameters of call expressions, either for some
ghost arguments or for the `by` proof. I looked at various
possibilities:
  * ``locals`call.i``
  * ``locals.call`i``
  * ``locals`call`i``
  * ```locals``i```
  * ```call`i```

We decided to go with the last one for the following reason
- It's as if a sub-call was creating a separate object, and ì` is one of
its fields. Even if it's allocated on the stack. Internally, we still
resolve it to a call.
- It's used only in the context of a method call with the same name, so
it's not ambiguous (double backtick was ambiguous in that it might refer
to the stack below or above)


### How has this been tested?
I have three test files
- One to test everything when everything passes verification
- One to test verification errors
- One to test resolution errors

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>