Fix: Dafny Rust runtime no longer checks if RC increased by one when concurrent (#6248)
Fixes #6247

### What was changed?
- Removed a check that the reference count was increased by 1 when the
dafny runtime is used concurrently. This check was performed only to
verify the compiler soundness, in that it would actually increase the
resource count and not ignore the sentence in the run-time.

### How has this been tested?
- A test performed only in concurrent mode makes a lot of copies of an
object. That test fails if the line that checks reference count is
activated, and no segfault happens if the line that checks reference
count is omitted.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>

---------

Co-authored-by: Robin Salkeld <salkeldr@amazon.com>