Perform reads checks on ensures of bodiless functions (#6264)
### What was changed?
<!-- Is this a user-visible change? Remember to update RELEASE_NOTES.md
-->

- Add reads checks on ensures body for functions without a body. Closes
issue #5822
- Fix the test `comp/replaceables/complex/user.dfy`: this new feature
showed that the specification for these external functions was wrong.
For one of the function, a missing `reads` clause (flagged by this new
feature) allowed to derive `false`. The other function's spec was
implying that a **_function_** was allocating a new instance (namely of
`Option<Int32>` where `Int32` is an extern class), which is forbidden.
So we fix these specs by changing them to be methods instead.

### How has this been tested?
<!-- Tests can be added to
`Source/IntegrationTests/TestFiles/LitTests/LitTest/` or to
`Source/*.Test/…` and run with `dotnet test` -->

Add a test to show this new check, update an existing test that includes
such a bodiless function with reading ensures and without reads clause.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>

---------

Co-authored-by: Samuel Chassot <samuuu@amazon.com>