Function deserialization (#6269)
### What was changed?
Mimic the logic of adding `ArrowType` in the `SystemModuleModifiers`
when deserializing a function. This is similar to what is done during
parsing, see line 2015 of `Dafny.atg`

### How has this been tested?
Tried to extend the file `inputFormatAnything.dfy` but the change I
could not make it fail without my changes.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>