fix: Escape string literals in error messages (#6273)
This PR fixes a crash where a string literal in the input was being
passed to a string format method that interpreted the string as a format
string. Characters like unmatched `{`'s in that string had then caused a
crash when error messages were printed.


<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>