Fix: Shadowing of output variables in methods in Dafny-to-Rust (#6278)
Fixes #6261 

The root cause of the bug was that shadowed variables are never
accessible in Dafny anymore except for one case: Output variables,
through the return statement.
Indeed, a return statement with an expression assigns the result of the
expression to the (unique) output variable, and then returns. Even in
the presence of shadowing when another local variable overrides the
output variable, the (outer) output variable is correctly assigned in
the compiler.
I had not considered this case when considering enabling shadowing.

### What was changed?
- Whether a variable's name results in the shadowable name instead of
the unique name is now more constrained: it requires that the variable
name is different from any output variables.

### How has this been tested?
- Two tests, including one with a structure close to #6261 with the
elephant operator, were failing before this PR and now are passing.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>